Prevent AttributeError in MethodPredicate
Pyramid orders views based on the number and kind of predicates. In the
test, the batched view was always appearing first, but in some
real-world cases, it would not. When a method predicate ran, it would
try to get the 'rpc_method' attribute from the request and fail, raising
an AttributeError.

Fixes #39.